Use correct include paths for internal Bayeux
Bayeux's submodules are not aware of the top level bayeux directory
for headers, so this needs to be added to the include path so internal
includes work (e.g. #include "datatools/foo.h").

Update CMake scripts to add the extra include/bayeux directory, and
link Falaise executables to the transparent Bayeux_LIBRARIES variable.