Implement basic structure of flreconstruct
Following (roughly) design of flsimulate, provide basic implementation
of flreconstruct. Hand off main function to an implementation function
calling out to a command line dialog for user configuration, passing
the result of this parsing to a function which constructs and runs the
pipeline.

Do not use dpp_driver class as this does not provide a usable interface.