Optionally allow LaTeX generation
Provide advanced option to enable LaTeX generation by Doxygen.
Default to OFF, and only generate the TeX files, do not compile them

Explicitly order md files in doxygen config so that they appear in
logical order in the Related Pages in HTML and in the main LaTeX
manual.

Provide line breaks in Installation Guide to split long command lines
so these do not spill over in the LaTeX output.