Describe Cadfael bootstrap with custom cmake
Cadfael's bootstrapping script assumes CMake to be located in /usr/bin.
If this is not the case, the script will fail if run directly.

Describe how to run bootstrapping with a custom location for cmake. This
covers the cases of no /usr/bin/cmake, and also that of overriding the
system cmake.