Allow loading of custom modules in flreconstruct
Allow user's pipeline script to contain a section named 
"flreconstruct.plugins". This section should contain information on
the shared libraries to be loaded. Recognise a single main string key
"plugins" that holds a list of ids of shared libraries to be loaded.
Each id can have subkeys in the form "id.SUBKEY" where SUBKEY is
any of those recognised by datatools::library_loader. Use these
subkeys to generate a suitable multi_properties object for configuring
the datatools::library_loader instance of flreconstruct. Always
autoload libraries.

For example, a section

[name="flreconstruct.plugins" type=""]
plugins : string[2] = "z" "gsl"

would try to load "libz.so" and "libgsl.so" from the default paths of
the dynamic loader (including the CWD and paths in LD_LIBRARY_PATH).

The ids in "plugins" are just that, so we could equally write

[name="flreconstruct.plugins" type=""]
plugins : string[2] = "foo" "bar"
foo.filename : string = "libz.so"
bar.filename : string = "libgsl.so"

for the same effect.