Fix warnings in BRErrorAsString
Switch based control with return statements in each case results in
a "control reaches end of non-void function" compiler warning (Though
this is minor as assertion in default block prevents a fall off the end
issue).

Replace direct returns with setting of local string, returning this if
switch statement completes.