Add OS X support to examples
Update example CMake scripts to add link flags for symbol dynamic
lookup when builsing od Apple platforms. This is required to
prevent undefined symbol errors because we rely on flreconstruct
to provide the needed symbols of loadable modules.

Provide documentation on the CMake command used and brief explanation
of why it is needed.