Import root conversion module
Import code from Kieran and Yorck for the "Things2Root" module.
Update module to use geometry service as implemented in flreconstruct.
Note that only geiger data is handled as yet. 

Update CMake build script to install module in a special modules 
directory.

Point CMake script in modules/ subdirectory to buildtime location
of Bayeux so that modules use this when building. This permits
modules to be developed in isolation and later imported into
the mainline of Falaise easily.