Add example of service access/usage
Provide an example of accessing services in a custom module for 
flreconstruct. Demonstrate using the core geometry service, accessing
this and storing it in the module via the module's initialize method.

Provide Doxygen Markdown page describing services and a walkthrough of
the example.

Update doxygen buildscripts and frontpage to incorporate the page in
the overall guide.