Add some comments as a hint for possible future issue
When a program is using several output ROOT TFile objects with several
associated TTree objects,  we may face some issue  when trees' entries
are stored in a file.  Particularly, some alternately usage of 2 trees
each associated with  2 different output files may lead  to some badly
redirected  output (i.e.   the wrong  tree stored  in the  wrong file)
because the default  Root global scope is  changing.  Using explicitly
the TTree::SetDirectory(TFile*) method should force a given tree to be
definitively  attached to  a unique  TFile, preventing  TFile swapping
issue due to implicit behaviour of  Root with the TFile::cd() that may
chnage thecurrent directory scope.

For now, a single comment in added in the code where it is expected to
be useful for a future fix.  Note that this technique has been used in
the Bayeux/brio library.