Port the TrackFit library within Falaise
The trackfit C++ library  has been created a long time  ago to be used
within the  NEMO-3 NAT++ off-line  software suite  as well as  for the
SuperNEMO project.   It is  an independant  library with  an interface
that can be used easily with NEMO-3 and SuperNEMO data:

See the original code at: https://nemo.lpc-caen.in2p3.fr/svn/trackfit.

This  work is  a port  of trackfit  (renamed 'TrackFit')  that is  now
integrated in the Falaise library.

TrackFit provides two algorithms:
  * A fit manager  for a linear segment that  reprensents the straight
trajectory of a  charged particle in a NEMO-like  detector with Geiger
hits.
  * A  fit manager  for a  helix segment  that reprensents  the curved
trajectory of a  charged particle in a NEMO-like  detector with Geiger
hits.

A TrackFit driver class and  associated data processing module will be
provided  within  Falaise in  order  to  build candidate  trajectories
associated to clusters of Geiger hits.