Try hashtag reference to link markdown pages
Technique to link/reference Markdown pages in Doxygen appears to change
between 1.8.X series.

Update Markdown page for flsimulate to using "hashtag" style of
referencing. Add hashtag after primary title of flsimulate usage guide
and use this in @ref links on main page.

This is an initial test because of the need to trial on several platforms.
Confirmed to work on Linux/Doxygen 1.8.1.2.