Document use of mock calibration in flreconstruct
Add Markdown page to describe current state of mock digitization and
calibration. Note nature of mocking and lines of future development
and customization.

Provide example flreconstruct pipeline script demonstrating use
of mocking modules as a preprocessing step of the pipeline

Begin description of how to develop new models, emphasize need to
not overwrite official data both for clarity and to help comparison
and validation.