update resource files
The configuration files  for the Geant4 manager and  the embedded step
hit   processors   are    modified   for   'snemo::demonstrator'   and
'snemo::tracker_commissioning' setups in such a  way the track Ids are
recorded  within  the truth  step  hits,  together  will a  few  other
properties that  could be useful for  users that plan to  validate the
tracking  algorithms.  The  "gg.hit_processor" object  now stores  the
track information  (track Id and  parent track Ids) associated  to any
Geiger hit collected in the "gg"  collection of truth Geiger hits.  It
also computes and stores  the 'minimum_approach_position' to the anode
wire of any truth Geant4 step.

Note also  that as mctools now supports the configuration of  the step
hit  processor   factory  using   several  files,   the  configuration
parameters of the hit processors are splitted in two files:

  * one file  is for the  'official' sensitive detectors  that provide
the "calo",  "xcalo", "gveto" and  "gg" collections  of MC hits,

  * the  other  file  is  for sensitive  detectors  that  are  created
'on-the-fly' in order to feed the "__visu.tracks" collection of Geant4
step hits  from various  parts of  the detector.  This "__visu.tracks"
collection  is  supposed to  be  used  for visualization,  validation,
debug... and not for some realistic processing of the simulated data.

In the future, it should be possible to deactivate the processors from
the  second file  in order  to  generate only  hits with  experimental
meaning and  thus eliminate  the huge "__visu.tracks"  collection from
the simulation output.