first implementation of specialized calorimeter step hit processor for SuperNEMO
This is based on S. Calvez's work and actually, only implement
calorimeter locator for the main wall. An important point, is that if
the localization fails then we get back to the default algorithm located
in mctools library. In particular, this solves a problem when a gamma
interact deeply inside calorimeter block : calo_locator only look for
interaction within the front part of the calorimeter block (thickness of
31mm). Thus if we do not go back to default calorimeter step hit
processor, the Geant4 step hit is not properly clusterized and we get a
warning.