reset totally the geom_id
Only invalidating the geom_id kept the address depth which is bad when
we successively call different locators : for instance, when calling
calo_locator, the geom_id gets a depth of 5 compatible with xcalo
geom_id but incompatible with gveto geom_id (depth = 4). So calling
first calo_locator within calorimeter_step_hit_processor first sets the
depth to 5 even if the hit is not inside main wall calorimeter. Then
looking within gveto_locator, the depth was wrong and thus gveto
localization was always wrong even if the hit was actually well located.