adding a new module for processing 'tracker_trajectory_data'
The 'charged_particle_tracking' module uses 'tracker_trajectory'
i.e. the results of tracker cluster fit to reconstruct physical
quantities given the SuperNEMO detector geometry. It determines the
electric charge value, it extrapolates the vertices and it associates
trajectory to calorimeter hits: for each trajectory there might be a new
'particle_track' object which is the physical interpretation of the
fitted trajectory.