Implement output profiles from the simulation config (demonstrator)
Bayeux/mctools now supports *output profiles* that enable
the user to tune the level of details he/she want to
record from the Geant4 based simulation. The main effect
is to store additional collection of MC truth hits collected
from some volumes of interest but not *real* sensitive detectors.
This allows the visualization of true tracks in addition to
the only recording of hits in the various detectors (geiger cells
and scintillator blocks).

Some new properties are now defined in the configuration files
of the geant4 engine.