improve vertices extrapolation
The extrapolation of track vertices is now done within one unique method
that treat in the same way calorimeter vertices and source foil
vertices. A list of all vertices are thus computed and only the 2 "best"
one i.e. the ones that minimize the track path are kept. This avoids
some strange track shape such as a track crossing the foil or
interacting behind x-wall.