clarify class member name
The association particle track to calorimeter means association to a
triggered calorimeter block i.e. a calorimeter hit. This is slightly
different to association with a calorimeter block that may have not
triggered. The commit calrifies this terminology since the calorimeter
information retrieved by *_associated_calorimeter_hits() refers to
calorimeter hits.