Experimental support for ROOT output
Allow user to switch between BRIO and ROOT output based on the extension
of the output file they supply.

If output file is supplied and ends in ".root", create the output
module as an instance of Things2Root. Configure this instance to write
to the given file and use the configured services. For all other files
use BRIO format as before.

NOTE: The use of ROOT output *does* result in a segfault on linux
on the first event on the call to the fill of the ttree. This does
*not* happen if Things2Root is used via a config file. It is
currently not clear why this happens.

Update build script to #include and link directly to Things2Root
plugin. This does not affect behaviour on Linux, but is only a
temporary solution in order to try and resolve the above issue
with ROOT output.