enable/disable primary particle at the level of genbb::primary particles
If genbb primary particles have been generated with generation id
support, then it is possible to enable/disable tracks that belongs to
the same track id as teh primary. Visually it may be interesting to know
where a primary track goes especially when several primaries are the
same say gammas for example.