fix calo locator when returning neighbors ids
Main wall block are described as two scintillators : a front step and a
extruded back block. The get_neighbours_ids function returns a list of
geom_id arround a given calorimeter block but, so far, only the front
step was considered. This fix sets the 'part' address of the returned
geom_id as 'any' so it can be either the front or the back of the
block. Consequently, there is no more matching problem when comparing
the list of geom_ids with another calorimeter block.