show recalibrated tracker hits
For delayed tracks such as alpha, the Geiger hits can not be calibrated
due to large anodic time. After the fitting process especially when
using trackfit plugin, the fitting process can recalculate the start
time for these delayed tracks by adding it to the set of fit
parameters. It is then possible given the new t0, to show the new
recalibrated geiger hits and to compare it with the simulated drift
distance for instance.