add new module
GammaClustering module is a SuperNEMO module that not only clusterizes
calorimeter hits to identfy gamma particles but also, tracks gamma
clusters based on Time-Of-Flight measurement. This module is then a mix
between a pure gamma clustering approach and the gamma tracking
alogorithm proposed in GammaTracking modules repository. By combining
the advantages and features of both algorithms, it increases its
efficiency.