add alpha finder driver
Alpha finder driver tries to look within unclustered hits and unfitted
clusters to find short alpha delayed tracks. By short it means, alpha
with less than 3 Geiger cells associated otherwise the trackfit just
fits them with a line fit (and also deduced the time delay wrt to prompt
track).

The basics behind the algorithm is to search for delayed Geiger hits
closed to prompt tracks. The "proximity" is all configurable and it can
be changed by the user when setting up the driver.