make default drawer more generic
The default drawer is called when no dedicated drawer has been
created (such as SuperNEMO, for instance). The following commit makes the
representation of calorimeter hits from Monte-Carlo more generic in the
way that no category hits need to be provided : all the hit categories
are retrieved and assumed to be calroimetric information. Maybe we will
need more refinement in future to be sure the step hit is calorimetric.