remove ambiguity when a track is associated to 2 calorimeter hits
This is a first try to remove ambiguity when a particle track is
actually associated to 2 neighbour calorimeter hits. We basically look
if the calorimeter hits are neighbours and in such case, we onl consider
the calorimeter hit with the shortest distance to vertex and not already
associated. To be tested ...