prevent double counting of non associated calorimeter
When no tracker trajectory is found (pure gamma event for instance), all
the calorimeter in the event are added to the list of non associated
calorimeter hits at the beginning of the process. But at the end of the
process, the calorimeters hits are again added to the list of non
associated calorimeter hits. The commit prevents double addition of
calorimeter hits by ensuring that no non associated calorimeter hits
have been already added.