add 'isolated' tag for non associated calorimeter hits
Non associated calorimeter hits do not mean no geiger cells are close
enough to the calorimeter hit. For instance, a sole geiger cell is not a
tracker cluster and thus no tracker tracjectory and no calorimeter
association. So, an isolated calorimeter hits is a hit without any
Geiger cell in front. This tag allows the gamma clustering/trakcing
algorithms to only perform calculations on this kinf of calorimeters.