change the interface of gamma tracker/clusterizer algorithms
This major change in the interface of gamma builder algorithm allows the
base_gamma_builder to select which calorimeter hits go to the gamma
tracker/clusterizer algorithm. This avoid the duplication of code when
selecting hits for differents drivers/algorithms. This is mainly due to
additionnal information for the calorimeter block coming from Charged
Particle Tracking module.