Fix keyword redefine warnings
ROOT5's generated dictionary sources redefine accessor keywords
to assist with reflection. This is expected and correct
behaviour, but Clang will warn about it.

On Clang only, add additional compile flag for the generated
dictionary source to suppress this warning globally for that
file. Done like this because the source is generated.