Add documentation on usefulness of '-v' flag
Development of trunk Falaise/Bayeux against a software stack that 
contained an older Bayeux caused some odd compiler errors, traced
back to a reordering of include paths due to use of -isystem.

Tracking and resolving the issue was assisted by the use of the
'-v' flag to gcc/clang which prints out complete search paths
and much other useful info. Add a note the the Developing Falaise
docs on the usefulness of this flag.