Add explicit seeding support in flsimulate
A new switch is available from flsimulate in order to load an
explicit set of seeds for the embedded PRNGs. The new option
takes the name of a file which contains the explicit values
for 4 seeds used by the Geant4 simulation manager (class
'mygsl::prng_seed_manager'). It is recommended to used the 'bxg4_seeds'
application to generate the file which contains the seeds.

Example: Launch 3 statistically independant runs

  $ bxg4_seeds --number-of-runs 3 --run-start 0 --pattern "mc_seeds_%n.conf"
  $ flsimulate --input-seeds "mc_seeds_0.conf" --output-file "mc_out_0.brio"
  $ flsimulate --input-seeds "mc_seeds_1.conf" --output-file "mc_out_1.brio"
  $ flsimulate --input-seeds "mc_seeds_2.conf" --output-file "mc_out_2.brio"