Update resource files
The   'resources/config/snemo/demonstrator/experiment/1.0/'  directory
contains the  configuration files used  to modelize the  components of
the experimental setup in the Vire library and tools.

This first set of files enables the description of:

- the devices that  are part of the detector  itself (optical modules,
  geiger  drift  cells)
- the electronics  components  (racks,  crates, boards, channels...)

For each of these components (devices), we describe its placement into
the conventional hierarchy of the full assembly, we define a numbering
scheme (mapping), and we describe its interface (ports) to the outside
world (pins, connector, plugs...).

Such tools make possible to implement high level functionalites:

- represents the cabling between hardware and electronics devices
  (port to port links),
- generate conventional tables/word list,
- geenrate views of the experimental assembly...

It  is a  preliminary  work for  future use.   No  Vire dependency  is
introduced so far from Falaise.