Promote flsimulate_next candidate to production flsimulate
FLsimulate_next  is  promoted  to   the  new  'fLsimulate'  production
application  and  the  base  tool for  MCC1.0 (NemoDocDB-doc-4262-v1).

Former  FLsimulate  is preserved as the 'fLsimulate_legacy' application.

FLsimulate's interface  has changed.  It  now uses an input  script to
configure the full simulation engine, as proposed by Ben and similarly
to  the technique  used  in FLreconstruct.   It  supports natively  an
embedded variant service to allow the user to tweak some parameters of
interest published by  the variant system (geometry,  vertex and decay
generators, specific Monte Carlo options...).

Support for metadata has been enriched and refactored in a more strict
approach. Metadata should  be saved in a external  file and optionally
embedded  in the  output data.  They are  re-usable in  other contexts
(FLreconstruct, TBD)

This release proposes to use the new SuperNEMO demonstrator simulation
setup 2.1 and is compatible with the last geometry model (4.0) as well
as  the last  version of  the  vertex and  decays generation  managers
(resp. 4.1 and 1.2) that have been enhanced during the last months.

Documentation is updated in: documentation/flsimulate/FLSimulate.md.
An example is provided in: resources/examples/flsimulate/ex01.