Fix flvisualize for Issue #5
Installs of flvisualize cannot switch between 2d views. Traced
to ROOT 6 dictionary generation/use requiring runtime access to
headers used. Rootcling adds #include directives to the dictionary
"payload" and if these are supplied to rootcling as absolute paths,
one will get "#include /abs/path/to/build/time/location.h". The
installed dictionary will then not be able to use the header in
installs - not picked up in testing because the build time
directory was still present when running! It is not possible
to supply additional runtime include paths to ROOT to workaround
this because of the use of an absolute #include.

The absolute paths are supplied to rootcling by the ROOT_GENERATE_DICTIONARY
function supplied by Root, even if relative paths are supplied. This
is due to an underlying assumption in the function that headers are
stored under an "inc" subdirectory. It is not possible to workaround
this assumption nor supply paths to strip from supplied header paths.

To fix this issue for flvisualize reorganize the dictionary generation
and install as:

- Move header tree from include/ to source/inc/ to provide the
  root path assumed by ROOT_GENERATE_DICTIONARY.
- Install dictionary headers alongside EventBrowser {dylib,so}/pcm
- Prepend that location to ROOT_INCLUDE_PATH on flvisualize
  startup, using self-location of falaise.

The use of the ROOT_INCLUDE_PATH environment variable appears to be
required. The seemingly equivalent call to TSystem's AddIncludePath
member function still results in Cling failures.