[issue #8] Fix missing helix fit model
Default pipeline  lacks helix fit  for tracking.  Now both  "line" and
"helix" ft models are performed on charged particle tracks, regardless
the magnetic field is activated or not.

A  better fix  would  use  variant to  activate  helix  fit only  when
magnetic  field  is set  from  the  geometry  section of  the  variant
profile.   However,  the  Bayeux implementation  of  such  conditional
variant  is not  satisfactory and  may require  a fix.   This will  be
addressed in Bayeux itself.