Fix ROOT palette color
Setting ROOT style at the level of style_manager was actually not working. With
ROOT version lower than 6.04, the problem as actually hidden since the rainbow
palette (with 100 colors) was still the default one. In version 6.04, new
palettes were added and the rainbow palette was not anymore the default.

The ROOT style has to be done when the TCanvas are actually created by the
pad_embbeded_viewer i.e. when the TRootEmbbededCanvas is not only created and
instantiated but also used. The method update_detector is maybe not the best
place but for some obscure reasons, the only place where ROOT styles are
applied.