Add `clang-format` in guide
Create section in Getting Started to list additional packages that
developers should/must have setup to contribute to Falaise. Add
`clang-format` here, with guides for brew or system install and
setup.

Append running of clang-format to use of style guide.

TODO: For simplicity and clarity, review creation of a dedicated
page in documentation for "additional developer tools". We will
be adding more.