Add event header bank in flsimulate output
Event header based cuts, as proposed in PR #35 required the presence
of an "EH" event header bank in each event.

Provide a preliminary implementation of this in flsimulate. The
bank will propagate through flreconstruct/flvisualize and be
available for use via cuts or other processing.

As run number is a meaningless number in flsimulate, mark it as
"any" so that it comparisons can ignore it.