CMake tooling to run clang-format
Whilst it is recommended that clang-format be run via the user's
editor or IDE, it is also useful to allow the build to run it,
e.g. "make format". This can be helpful to less experienced users
and also when reviewing pull requests or preparing releases.

Add a new CMake module implementing a function `target_clang_format`
that creates a target to run clang-format over the source files of
a supplied target. Only create the formatting target if a clang-format
executable exists, searching for this when the module is included.

Add call to target_clang_format for each core library/executable.
Confirm that this runs, but do not commit formatting changes yet.