Documentation: Create script to prepare gh-pages
Doxygen HTML output files can contain absolute/relative paths to:

- Bayeux tagfiles
- Bayeux root dir for HTML docs
- Project Source/Build dir

Direct hosting these files on the Falaise gh-pages branch will thus
lead to broken links and cross-references.

Provide a cmake-configured script that a release manager can run
to strip/replace these paths from Doxygen HTML. Output processed
files to a clean directory in the build root. The directory contents
can be copied to the root dir of the gh-pages branch ready for
hosting.

The script is not integrated in the build system yet, as it is not
for general use.