Implement Git based build version tracking
Delete obsolete LPCSVNUtilities module, replacing with
GetGitRevisionDescription modules from rpavlik's cmake-modules:

https://github.com/rpavlik/cmake-modules/blob/master/GetGitRevisionDescription.cmake

Used as it ensures that cmake will rerun when the HEAD/refs/index changes,
ensuring the hash exported to the build is up-to-date.

Retain core versioning in project() call, appending tags to this when
in a git working copy as follows:

- Add "-g<First 8 characters of current HEAD hash>"
- Add `-dirty` if the working copy includes uncommited changes

Export Hash and "dirty" state as string and symbol respectively in
Falaise's version header. Add methods to return commit hash and
state. Retain old SVN "revision" interface for backward compatibility,
marking as deprecated and giving negative revision number. Implement
tests to confirm compiletime/runtime versions/tags are the same.

Add build's commit hash and state to output of `--version` calls to
`flsimulate` and `flreconstruct`. Also add this info to the metadata
entries in flreconstruct and flsimulate's output under the
`falaise.version` section.