Modernize tracker_trajectory_solution
Key change is marking of get_auxiliaries member functions as deprecated.
They are only used in the EventBrowser, and only then to store View
related information. This is completly unrelated to the job of the
data model, but needs retention for now due to the tightly coupled
nature of the design and programming style.

Protect use in EventBrowser with pragmas to suppress warnings in
known locations but flag up others.