Demonstrate simplified tree_dump
Whilst data is not directly designed to be human readable, most classes
implement a "tree_dump" member function (it should be operator<<, but
that's by the by...). When dumping collections, a contrived algorithm
is used to implement the pattern of "do one thing for all-but-last
element, something else for last". Typically, this is to change prefixes
for the last line.

Demonstrate proper use of pattern via C++ standard algorithms as a
temporary measure.