Minimal bounded integer implementation and test
The geometry and reconstruction algorithms in Falaise have many places
where a limited range of integers is used to address "things", for
example tracker cell indices. Having values outside these ranges is
an error, and whilst these are tracked, the resulting checks are
bolierplated throughout the code.

Provide a basic "bounded_integer" type to represent a range
[lower,upper] of integral values. Constructing or assigning with a
value outside of this range results in an out of range exception.

Implement minimal unit tests of this behaviour