Always store metadata in output sim file
Remove flsimulates's ability to write metadata to anything other then
the output data file. This avoids ambiguity and complexity in storing
and understanding this data later on.

Replace separate file output with new command line program
`brio_file_dumper`. This dumps the metadata in an input brio file
to stdout in multiproperties format, allowing redirection to file if
required.