Make commit hash/state symbols private
As state of git changes frequently with development, so does the
interface of version.h. This leads to frequent recompiles of tests
when only the hash or state has changed.

Because hash/state are really metadata about the compiled state,
make the preprocessor symbols only visible in version.cc. Both
the hash and state can be queried at runtime via the existing class
functions.