Free function to add data products to event with checks
An identified use case in Falaise modules is to add data products
to the event record under a given key if and only if the key does
not exist in the record. Whilst datatools::things's add member function
does check this, it happens relatively late and with a not so helpful
exception.

Provide a free template function to allow modules to add a data product
to an event record with:

- Fast runtime check on key existence with throw of dedicated existing key
  exception in case key exists.
- Compile time check that type of data product meets requirements for
  storage in the event record.

Add the compile time check to the existing getOrAddToEvent function.