Convenience member function for self default parameters
A common pattern identified in reconstruction modules use of
property_set is "self defaulting":

```
// Usually in constructor
double a = 3.14;

a = ps.get<double>("key_for_a", a);
```

i.e. the parameter "key_for_a" is not required, and should take
'a's existing value when the key is not present.

Provide a clearer interface for this use case "assign_if", that
takes the variable to be assigned to by non-const reference.
Assign the value held at the supplied key if it exists and of
the matching type. Throw on mismatched types, leave variable
unchanged if key is missing.