Require falaise::path type for mapped_magnetic_field parameter
As reported in #198, flsimulate will fail to run with the mapped
magnetic field variant, throwing an exception:

```
flsimulate : Setup/run of simulation threw exception
value at 'map_file' is not of requested type
```

Traced to the `map_file` parameter being correctly declared in the
config file as `string as path`, but extracted as `string` in the
mapped_magnetic_field initialize() member function.

Extract parameter explicity as type `falaise::path` for correct
validation.