Initial clang-tidy config for Falaise
Allows clang-tidy to be run with a standard set of checks without
needing to specify on the command line. All checks are turned off apart
from performance, readability and modernize. In addition, two additional
checks are turned off:

- modernize-use-trailing-return-type
  - May lead to a bit too much confusion at present
- readability-magic-numbers
  - Do have lots of magic numbers (unamed numerical constants) in code,
    but major refactor to remove them. Turn off for now to reduce noise.

If clang-tidy is installed on the system, then all that's required to
use is to run CMake as:

```
$ cmake -DCMAKE_CXX_CLANG_TIDY=/path/to/clang-tidy` <otherargs>
```

when building, clang-tidy will check and report on detected "violations"
and print where these were found and the recommended fix.

Fixes can in many cases be applied automatically by configuring with:

```
$ cmake -DCMAKE_CXX_CLANG_TIDY=/path/to/clang-tidy;-fix` <otherargs>
```

Note that this can result in build errors if the fix is on a
particularly badly structure block of code. It may also require several
rebuilds to fully resolve fixes (as fixes may result in other fixes
being opened up).