Implement a module wrapper for event filtering
Discussion identified the need for "cuts" in Falaise pipelines to
select/reject events for onward processing. This would in most cases
be pure keep/drop, but the case of wanting to continue both branches
along different processing pipelines should be considered, e.g.:

                     - Path on Acceptance ->
                    /
Input -> A -> Filter
                    \
                     - Path on Rejection ->

Provide a new "filter" class inheriting from dpp::base_module to
implement this concept. It wraps a filter algorithm supplied by
the implementer, handling the input configuration and the direction
to the acceptance-rejection paths appropriately.