Implement "black hole" filter module
Many use cases for event filtering require no further processing on
or interest in events that filters reject. This includes not writing
rejected events to file.

Provide a builtin "black_hole_module" that does no processing and
always returns a PROCESS_STOP status to indicate to the pipeline
that no further processing or output of the event should take place.